Implement saving a gigg

+++ Gighub/ViewModels/GigFormViewModel.cs
@@ -1,8 +1,5 @@
 ﻿using Gighub.Models;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 
 namespace Gighub.ViewModels
 {
@@ -11,7 +8,7 @@ namespace Gighub.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time{ get; set; }
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
         public IEnumerable<Genre> Genres { get; set; }
     }
 }
